# backend/generated/popcornhub_api/controllers/default_controller.py
import connexion
from typing import Dict
from typing import Tuple
from typing import Union

from popcornhub_api.models.ai_chat_post_request import AiChatPostRequest  # noqa: E501
from popcornhub_api.models.chat_response import ChatResponse  # noqa: E501
from popcornhub_api.models.error import Error  # noqa: E501
from popcornhub_api.models.health_get200_response import HealthGet200Response  # noqa: E501
from popcornhub_api.models.movie import Movie  # noqa: E501
from popcornhub_api.models.movie_suggestion import MovieSuggestion  # noqa: E501
from popcornhub_api.models.watchlist_delete200_response import WatchlistDelete200Response  # noqa: E501
from popcornhub_api.models.watchlist_post200_response import WatchlistPost200Response  # noqa: E501
from popcornhub_api.models.watchlist_request import WatchlistRequest  # noqa: E501
from popcornhub_api import util


def ai_chat_post(body):  # noqa: E501
    """AI movie chat

    Get movie recommendations through AI chat # noqa: E501

    :param ai_chat_post_request: 
    :type ai_chat_post_request: dict | bytes

    :rtype: Union[ChatResponse, Tuple[ChatResponse, int], Tuple[ChatResponse, int, Dict[str, str]]
    """
    ai_chat_post_request = body
    if connexion.request.is_json:
        ai_chat_post_request = AiChatPostRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def health_get():  # noqa: E501
    """Health check

    Check if the API is running # noqa: E501


    :rtype: Union[HealthGet200Response, Tuple[HealthGet200Response, int], Tuple[HealthGet200Response, int, Dict[str, str]]
    """
    return 'do some magic!'


def movies_get(q=None, genre=None, limit=None, offset=None):  # noqa: E501
    """Get movies

    Retrieve a list of movies with optional filtering # noqa: E501

    :param q: Search query for movie title or description
    :type q: str
    :param genre: Filter by genre
    :type genre: str
    :param limit: Maximum number of movies to return
    :type limit: int
    :param offset: Number of movies to skip
    :type offset: int

    :rtype: Union[List[Movie], Tuple[List[Movie], int], Tuple[List[Movie], int, Dict[str, str]]
    """
    return 'do some magic!'


def movies_movie_id_get(movie_id):  # noqa: E501
    """Get movie by ID

    Retrieve detailed information about a specific movie # noqa: E501

    :param movie_id: Movie ID
    :type movie_id: int

    :rtype: Union[Movie, Tuple[Movie, int], Tuple[Movie, int, Dict[str, str]]
    """
    return 'do some magic!'


def movies_recommendations_get(limit=None):  # noqa: E501
    """Get movie recommendations

    Retrieve personalized movie recommendations # noqa: E501

    :param limit: Maximum number of recommendations to return
    :type limit: int

    :rtype: Union[List[Movie], Tuple[List[Movie], int], Tuple[List[Movie], int, Dict[str, str]]
    """
    return 'do some magic!'


def movies_search_get(q, limit=None):  # noqa: E501
    """Search movies

    Search for movies by title, description, or other criteria # noqa: E501

    :param q: Search query
    :type q: str
    :param limit: Maximum number of results to return
    :type limit: int

    :rtype: Union[List[MovieSuggestion], Tuple[List[MovieSuggestion], int], Tuple[List[MovieSuggestion], int, Dict[str, str]]
    """
    return 'do some magic!'


def watchlist_delete(body):  # noqa: E501
    """Remove movie from watchlist

    Remove a movie from the current user&#39;s watchlist # noqa: E501

    :param watchlist_request: 
    :type watchlist_request: dict | bytes

    :rtype: Union[WatchlistDelete200Response, Tuple[WatchlistDelete200Response, int], Tuple[WatchlistDelete200Response, int, Dict[str, str]]
    """
    watchlist_request = body
    if connexion.request.is_json:
        watchlist_request = WatchlistRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'


def watchlist_get():  # noqa: E501
    """Get user watchlist

    Retrieve the current user&#39;s watchlist # noqa: E501


    :rtype: Union[List[int], Tuple[List[int], int], Tuple[List[int], int, Dict[str, str]]
    """
    return 'do some magic!'


def watchlist_post(body):  # noqa: E501
    """Add movie to watchlist

    Add a movie to the current user&#39;s watchlist # noqa: E501

    :param watchlist_request: 
    :type watchlist_request: dict | bytes

    :rtype: Union[WatchlistPost200Response, Tuple[WatchlistPost200Response, int], Tuple[WatchlistPost200Response, int, Dict[str, str]]
    """
    watchlist_request = body
    if connexion.request.is_json:
        watchlist_request = WatchlistRequest.from_dict(connexion.request.get_json())  # noqa: E501
    return 'do some magic!'