# backend/generated/popcornhub_api/models/movie_suggestion.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from popcornhub_api.models.base_model import Model
from popcornhub_api import util


class MovieSuggestion(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None, title=None, poster=None, year=None, rating=None):  # noqa: E501
        """MovieSuggestion - a model defined in OpenAPI

        :param id: The id of this MovieSuggestion.  # noqa: E501
        :type id: int
        :param title: The title of this MovieSuggestion.  # noqa: E501
        :type title: str
        :param poster: The poster of this MovieSuggestion.  # noqa: E501
        :type poster: str
        :param year: The year of this MovieSuggestion.  # noqa: E501
        :type year: int
        :param rating: The rating of this MovieSuggestion.  # noqa: E501
        :type rating: float
        """
        self.openapi_types = {
            'id': int,
            'title': str,
            'poster': str,
            'year': int,
            'rating': float
        }

        self.attribute_map = {
            'id': 'id',
            'title': 'title',
            'poster': 'poster',
            'year': 'year',
            'rating': 'rating'
        }

        self._id = id
        self._title = title
        self._poster = poster
        self._year = year
        self._rating = rating

    @classmethod
    def from_dict(cls, dikt) -> 'MovieSuggestion':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The MovieSuggestion of this MovieSuggestion.  # noqa: E501
        :rtype: MovieSuggestion
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this MovieSuggestion.


        :return: The id of this MovieSuggestion.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this MovieSuggestion.


        :param id: The id of this MovieSuggestion.
        :type id: int
        """

        self._id = id

    @property
    def title(self) -> str:
        """Gets the title of this MovieSuggestion.


        :return: The title of this MovieSuggestion.
        :rtype: str
        """
        return self._title

    @title.setter
    def title(self, title: str):
        """Sets the title of this MovieSuggestion.


        :param title: The title of this MovieSuggestion.
        :type title: str
        """

        self._title = title

    @property
    def poster(self) -> str:
        """Gets the poster of this MovieSuggestion.


        :return: The poster of this MovieSuggestion.
        :rtype: str
        """
        return self._poster

    @poster.setter
    def poster(self, poster: str):
        """Sets the poster of this MovieSuggestion.


        :param poster: The poster of this MovieSuggestion.
        :type poster: str
        """

        self._poster = poster

    @property
    def year(self) -> int:
        """Gets the year of this MovieSuggestion.


        :return: The year of this MovieSuggestion.
        :rtype: int
        """
        return self._year

    @year.setter
    def year(self, year: int):
        """Sets the year of this MovieSuggestion.


        :param year: The year of this MovieSuggestion.
        :type year: int
        """

        self._year = year

    @property
    def rating(self) -> float:
        """Gets the rating of this MovieSuggestion.


        :return: The rating of this MovieSuggestion.
        :rtype: float
        """
        return self._rating

    @rating.setter
    def rating(self, rating: float):
        """Sets the rating of this MovieSuggestion.


        :param rating: The rating of this MovieSuggestion.
        :type rating: float
        """

        self._rating = rating