# backend/generated/popcornhub_api/models/watchlist_request.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from popcornhub_api.models.base_model import Model
from popcornhub_api import util


class WatchlistRequest(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, movie_id=None):  # noqa: E501
        """WatchlistRequest - a model defined in OpenAPI

        :param movie_id: The movie_id of this WatchlistRequest.  # noqa: E501
        :type movie_id: int
        """
        self.openapi_types = {
            'movie_id': int
        }

        self.attribute_map = {
            'movie_id': 'movie_id'
        }

        self._movie_id = movie_id

    @classmethod
    def from_dict(cls, dikt) -> 'WatchlistRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The WatchlistRequest of this WatchlistRequest.  # noqa: E501
        :rtype: WatchlistRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def movie_id(self) -> int:
        """Gets the movie_id of this WatchlistRequest.


        :return: The movie_id of this WatchlistRequest.
        :rtype: int
        """
        return self._movie_id

    @movie_id.setter
    def movie_id(self, movie_id: int):
        """Sets the movie_id of this WatchlistRequest.


        :param movie_id: The movie_id of this WatchlistRequest.
        :type movie_id: int
        """
        if movie_id is None:
            raise ValueError("Invalid value for `movie_id`, must not be `None`")  # noqa: E501

        self._movie_id = movie_id